requirements.txt
flask
pytest
psycopg2-binary
numpy
Flask-restful